splitted the whole library into two modules ui and core
