+++ projects/decorator-driven-dynamic-form/src/lib/core/models/decorators/validation/sync/pattern.ts
