+++ projects/decorator-driven-dynamic-form/src/lib/ui/components/dynamic-form-input/dynamic-form-input.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { InputNode } from '../../models/types/inputs/input-node';
-import { InputTypes } from '../../models/types/inputs/input-types.enum';
+import { InputNode } from '../../../core/models/types/inputs/input-node';
+import { InputTypes } from '../../../core/models/types/inputs/input-types.enum';
+
 import { DataLoaderService } from '../../services/data-loader/data-loader.service';
 
 @Component({
