+++ projects/decorator-driven-dynamic-form/src/public-api.ts
@@ -7,37 +7,37 @@ export * from './lib/decorator-driven-dynamic-forms.module';
 // export * from './lib/views/dynamic-form/dynamic-form.component';
 // models
 // forms
-export * from './lib/models/decorators/forms/forms';
-export * from './lib/models/decorators/inputs/inputs';
+export * from './lib/core/models/decorators/forms/forms';
+export * from './lib/core/models/decorators/inputs/inputs';
 
 // actions
-export * from './lib/models/decorators/actions/reset';
-export * from './lib/models/decorators/actions/submit';
-export * from './lib/models/decorators/actions/button';
+export * from './lib/core/models/decorators/actions/reset';
+export * from './lib/core/models/decorators/actions/submit';
+export * from './lib/core/models/decorators/actions/button';
 
 // controls
 
 // validation
-export * from './lib/models/decorators/validation/sync/not-null';
-export * from './lib/models/decorators/validation/sync/required-true';
-export * from './lib/models/decorators/validation/sync/max';
-export * from './lib/models/decorators/validation/sync/min';
-export * from './lib/models/decorators/validation/sync/max-length';
-export * from './lib/models/decorators/validation/sync/min-length';
-export * from './lib/models/decorators/validation/sync/pattern';
-export * from './lib/models/decorators/validation/sync/email';
+export * from './lib/core/models/decorators/validation/sync/not-null';
+export * from './lib/core/models/decorators/validation/sync/required-true';
+export * from './lib/core/models/decorators/validation/sync/max';
+export * from './lib/core/models/decorators/validation/sync/min';
+export * from './lib/core/models/decorators/validation/sync/max-length';
+export * from './lib/core/models/decorators/validation/sync/min-length';
+export * from './lib/core/models/decorators/validation/sync/pattern';
+export * from './lib/core/models/decorators/validation/sync/email';
 
-export * from './lib/models/types/actions/actions';
-export * from './lib/models/types/inputs/input-types.enum';
+export * from './lib/core/models/types/actions/actions';
+export * from './lib/core/models/types/inputs/input-types.enum';
 
-export * from './lib/services/form-entity-processor/form-entity-processor.service';
+export * from './lib/core/services/form-entity-processor/form-entity-processor.service';
 
-export * from './lib/models/types/forms/form-update-strategy';
-export * from './lib/models/types/forms/form-value-transformer';
+export * from './lib/core/models/types/forms/form-update-strategy';
+export * from './lib/core/models/types/forms/form-value-transformer';
 
-export * from './lib/models/decorators/validation/cross/cross-validation';
+export * from './lib/core/models/decorators/validation/cross/cross-validation';
 
-export * from './lib/models/decorators/validation/async/async-validation';
+export * from './lib/core/models/decorators/validation/async/async-validation';
 
 // components
-export * from './lib/components/dynamic-form/dynamic-form.component';
+export * from './lib/ui/components/dynamic-form/dynamic-form.component';
